added validation

+++ Salon/Models/ServiceProductUsageModel.cs
@@ -14,7 +14,7 @@ namespace Salon.Models
         public int product_id { get; set; }
         public string serviceName { get; set; }
         public string product_name { get; set; }
-        public int unit_per_volume { get; set; }
+        public double unit_per_volume { get; set; }
 
         public double usage_amount { get; set; }
         public double total_usage_amount { get; set; }
